Refactor code

+++ ums-service-enrollment/src/main/java/com/uth/ums/enrollment/repository/StudentRepository.java
@@ -3,5 +3,6 @@ package com.uth.ums.enrollment.repository;
 import com.uth.ums.enrollment.model.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface StudentRepository extends JpaRepository<Student, Integer> {
+    // You can define custom query methods or use the ones provided by JpaRepository
 }
